requirements.txt
langchain==0.2.7
langchain_huggingface==0.0.3
python-dotenv==1.0.1
streamlit==1.12.0
